refactor: handling build warnings and improved swagger instances
