+++ cabal.project
@@ -47,7 +47,7 @@ source-repository-package
 source-repository-package
   type: git
   location: https://github.com/maestro-org/haskell-sdk
-  tag: 83a37d8e9fc8ee226522068b40c275bbb77b55f4
+  tag: bdab74cfffee5a11eccb68b9e2b8f8feba656950
   --sha256: TODO:
 
 -- Unfortunately, cardano-node 8.1.2 is constrained with plutus-ledger-api 1.5.0.0 and we would like at least 1.6.0.0.
