+++ geniusyield-server-lib/geniusyield-server-lib.cabal
@@ -104,7 +104,6 @@ library
     , ply-core
     , rio
     , servant-foreign
-    , servant-py
     , servant-server
     , servant-swagger
     , strict
